pyproject.toml
[tool.poetry]
name = "ilkin-umarov"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "3.12.6"
fastapi = "0.115.2"
uvicorn = "0.31.1"
python-jose = "3.3.0"
passlib = "1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
